2026/harsh-singh-multimodal-ai-eval/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
assemblyai>=0.30.0
requests>=2.31.0